# hw2/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: House App
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /house/{id}:
    x-swagger-router-controller: house
    get:
      operationId : getIt
      description: call GET function to get a house
      parameters:
        - name: id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetHouseResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      operationId: postIt
      description: call POST function to post a new house
      parameters:
        - name: id
          description: House ID
          type: string
          in: path
          required: true
        - name: model
          description: Model of house
          in: body
          required: true
          schema:
            $ref: "#/definitions/House"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      operationId : putIt
      description:  call PUT function to update a house
      parameters:
        - name: id
          description: House ID
          type: string
          in: path
          required: true
        - name: model
          description: model of house
          in: body
          required: true
          schema:
            $ref: "#/definitions/House"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      operationId : deleteIt
      description: delete a house
      parameters:
        - name: id
          description: House ID
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  GetHouseResponse:
    required:
      - id
      - model
      - year
    properties:
      id:
        type: string
      model:
        type: string
      year:
        type: number
  House:
    type: object
    properties:
      model:
        type: string
        description: model name of the house
      year:
        type: number
        description: year of the house
    required:
      - model
      - year
  GeneralResponse:
    type: object
    properties:
      success:
        type: string
      description:
        type: string
    required:
      - success
      - description
  ErrorResponse:
    type: object
    required:
      - message
      - success
    properties:
      message:
        type: string
      success:
        type: string